FIX Solr configuration and atom feed tests, stubs, template content links
